requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.7
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
joblib==1.5.2
numpy==2.3.4
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
ratelimit==2.2.1
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
whitenoise==6.11.0
xgboost==3.1.1
